+++ CHANGELOG.rst
@@ -11,6 +11,13 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 * WiP: Added an option the import pre-built quantum dots.
 
 
+0.8.7
+*****
+* Replaced ``print()`` calls with ``logger.warning()`` in all dye-related functions.
+* Added a recipe for calculating the synthetic accessibility score (SAS)
+  as adapted from `MolGAN <https://github.com/nicola-decao/MolGAN>`_.
+
+
 0.8.6
 *****
 * Added documentation for the new Nano-CAT ``multi_ligand_job()`` recipe.